remove stranded strace and move to task file
