+++ physionet-django/console/views.py
@@ -18,8 +18,6 @@ from django.utils import timezone
 from django.db import DatabaseError, transaction
 from django.db.models import Q, CharField, Value, IntegerField, F, functions
 from background_task import background
-from background_task.models import Task, task_failed, task_rescheduled
-from django.dispatch import receiver
 from django.contrib.sites.models import Site
 
 from notification.models import News
@@ -40,37 +38,6 @@ from console.tasks import associated_task, get_associated_tasks
 from django.conf import settings
 
 LOGGER = logging.getLogger(__name__)
-
-
-@receiver(task_rescheduled, sender=Task)
-def task_rescheduled_handler(sender, **kwargs):
-    """
-    Notify the admins when a task has failed and rescheduled
-    """
-    pdb.set_trace()
-    name = kwargs['task'].verbose_name
-    attempts = kwargs['task'].attempts
-    last_error = kwargs['task'].last_error
-    date_time = kwargs['task'].run_at
-    task_name = kwargs['task'].task_name
-    task_params = kwargs['task'].task_params
-    notification.task_rescheduled_notify(name, attempts, last_error,
-                                         date_time, task_name, task_params)
-
-
-@receiver(task_failed, sender=Task)
-def task_failed_handler(sender, **kwargs):
-    """
-    Notify the admins when a task has failed and removed from queue
-    """
-    name = kwargs['completed_task'].verbose_name
-    attempts = kwargs['completed_task'].attempts
-    last_error = kwargs['completed_task'].last_error
-    date_time = kwargs['completed_task'].failed_at
-    task_name = kwargs['completed_task'].task_name
-    task_params = kwargs['completed_task'].task_params
-    notification.task_failed_notify(name, attempts, last_error, date_time,
-                                    task_name, task_params)
 
 
 @associated_task(PublishedProject, 'pid')
